Move reusable auth related classes to service base module.
