+++ server/src/test/java/org/eclipse/hono/server/StandaloneTelemetryApiTest.java
@@ -19,7 +19,7 @@ import java.util.stream.IntStream;
 import org.apache.qpid.proton.amqp.Binary;
 import org.apache.qpid.proton.amqp.messaging.Data;
 import org.apache.qpid.proton.message.Message;
-import org.eclipse.hono.authentication.impl.AcceptAllPlainAuthenticationService;
+import org.eclipse.hono.authentication.impl.FileBasedAuthenticationService;
 import org.eclipse.hono.authorization.impl.InMemoryAuthorizationService;
 import org.eclipse.hono.client.HonoClient;
 import org.eclipse.hono.client.MessageSender;
@@ -99,7 +99,7 @@ public class StandaloneTelemetryApiTest {
 
         vertx.deployVerticle(registrationAdapter, registrationTracker.completer());
         vertx.deployVerticle(InMemoryAuthorizationService.class.getName(), authTracker.completer());
-        vertx.deployVerticle(AcceptAllPlainAuthenticationService.class.getName(), authenticationTracker.completer());
+        vertx.deployVerticle(FileBasedAuthenticationService.class.getName(), authenticationTracker.completer());
 
         CompositeFuture.all(registrationTracker, authTracker)
         .compose(r -> {
