+++ service-base/src/main/java/org/eclipse/hono/service/auth/Activity.java
@@ -9,7 +9,7 @@
  * Contributors:
  *    Bosch Software Innovations GmbH - initial API and implementation and initial documentation
  */
-package org.eclipse.hono.service.authorization;
+package org.eclipse.hono.service.auth;
 
 /**
  * Defines permissions that are required to access Hono service resources.
