+++ service-base/src/main/java/org/eclipse/hono/service/auth/AuthorizationConstants.java
@@ -9,7 +9,7 @@
  * Contributors:
  *    Bosch Software Innovations GmbH - initial creation
  */
-package org.eclipse.hono.service.authorization;
+package org.eclipse.hono.service.auth;
 
 import io.vertx.core.json.JsonObject;
